requirements.txt
# Additional test requirements go here
django<3.0  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.7.0  # https://github.com/joke2k/django-environ
django-allauth==0.45.0  # https://github.com/pennersr/django-allauth
django-countries==7.2.1 # https://github.com/SmileyChris/django-countries
django-treebeard==4.5.1 # https://github.com/django-treebeard/django-treebeard
djangorestframework==3.11.2  # https://github.com/encode/django-rest-framework
Pillow==7.1.1  # https://github.com/python-pillow/Pillow
django-filter==2.4.0 # https://github.com/carltongibson/django-filter
pyjwt==1.7.1 # https://github.com/jpadilla/pyjwt
Jinja2==3.0.2 # https://github.com/pallets/jinja
celery==4.4.2  # pyup: < 5.0  # https://github.com/celery/celery

django-filer==2.0.2
django-json-widget==1.0.1

django-sso-app
